requirements.txt
flask
python-dotenv
google-generativeai
pinecone
cohere
tiktoken
numpy
pypdf
gunicorn
